refactor(memory): simplify provenance key collections
